Moved component to common
